removed Rooms page and moved functionality to the homepage
